Delete per le visite correttamente implementata

+++ back-end/back-end-leopardi/src/main/java/leopardiproject/csd/repository/PrenotazioneVisitaRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import leopardiproject.csd.model.PrenotazioneVisita;
 
 @Repository
-public interface PrenotazioneVisitaRepository extends JpaRepository<PrenotazioneVisita, Integer> {
+public interface PrenotazioneVisitaRepository extends JpaRepository<PrenotazioneVisita, Long> {
 
 }
